move test utility to utils package
