+++ agent/src/test/java/com/teamscale/jacoco/agent/store/upload/delay/DelayedCommitDescriptorRetrievalTest.java
@@ -1,6 +1,7 @@
 package com.teamscale.jacoco.agent.store.upload.delay;
 
 import com.teamscale.jacoco.agent.git_properties.GitPropertiesLocator;
+import com.teamscale.jacoco.agent.utils.InMemoryStore;
 import org.junit.Test;
 
 import java.io.File;
